fix: move fn to ast_utils.py
